refactor: break out and refactor confetti component
